+++ hoomd/md/CommunicatorGridGPU.h
@@ -1,0 +1,87 @@
+/*
+Highly Optimized Object-oriented Many-particle Dynamics -- Blue Edition
+(HOOMD-blue) Open Source Software License Copyright 2009-2016 The Regents of
+the University of Michigan All rights reserved.
+
+HOOMD-blue may contain modifications ("Contributions") provided, and to which
+copyright is held, by various Contributors who have granted The Regents of the
+University of Michigan the right to modify and/or distribute such Contributions.
+
+You may redistribute, use, and create derivate works of HOOMD-blue, in source
+and binary forms, provided you abide by the following conditions:
+
+* Redistributions of source code must retain the above copyright notice, this
+list of conditions, and the following disclaimer both in the code and
+prominently in any materials provided with the distribution.
+
+* Redistributions in binary form must reproduce the above copyright notice, this
+list of conditions, and the following disclaimer in the documentation and/or
+other materials provided with the distribution.
+
+* All publications and presentations based on HOOMD-blue, including any reports
+or published results obtained, in whole or in part, with HOOMD-blue, will
+acknowledge its use according to the terms posted at the time of submission on:
+http://codeblue.umich.edu/hoomd-blue/citations.html
+
+* Any electronic documents citing HOOMD-Blue will link to the HOOMD-Blue website:
+http://codeblue.umich.edu/hoomd-blue/
+
+* Apart from the above required attributions, neither the name of the copyright
+holder nor the names of HOOMD-blue's contributors may be used to endorse or
+promote products derived from this software without specific prior written
+permission.
+
+Disclaimer
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDER AND CONTRIBUTORS ``AS IS'' AND
+ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, AND/OR ANY
+WARRANTIES THAT THIS SOFTWARE IS FREE OF INFRINGEMENT ARE DISCLAIMED.
+
+IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
+INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE
+OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
+ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+*/
+
+// Maintainer: jglaser
+
+
+#ifndef __COMMUNICATOR_GRID_GPU_H__
+#define __COMMUNICATOR_GRID_GPU_H__
+
+#ifdef ENABLE_CUDA
+#include "CommunicatorGrid.h"
+
+#ifdef ENABLE_MPI
+/*! Class to communicate the boundary layer of a regular grid (GPU version)
+ */
+template<typename T>
+class CommunicatorGridGPU : public CommunicatorGrid<T>
+    {
+    public:
+        //! Constructor
+        CommunicatorGridGPU(boost::shared_ptr<SystemDefinition> sysdef, uint3 dim,
+            uint3 embed, uint3 offset, bool add_outer_layer_to_inner);
+
+        //! Communicate grid
+        virtual void communicate(const GPUArray<T>& grid);
+
+    protected:
+        unsigned int m_n_unique_recv_cells;       //!< Number of unique receiving cells
+
+        //! Initialize grid communication
+        virtual void initGridCommGPU();
+
+    private:
+        GPUArray<unsigned int> m_cell_recv;       //!< Array of per-cell receive elements (multiple possible)
+        GPUArray<unsigned int> m_cell_recv_begin; //!< Begin of recv indices per cell
+        GPUArray<unsigned int> m_cell_recv_end;   //!< End of recv indices per cell
+    };
+
+#endif // ENABLE_MPI
+#endif // __COMMUNICATOR_GRID_GPU_H__
+#endif